making changes according to the comments of mentors
