+++ src/components/pages/dashboard-pages/HomeTeacher.jsx
@@ -31,7 +31,7 @@ const HomeTeacher = ({ profile, onNavigate, profileError }) => {
         </div>
         {profileError.message && <p>{profileError.message}</p>}
         <div className="flex flex-col w-3/4 p-2 items-center justify-center gap-4 bg-pureWhite ">
-          <h3 className="font-spartan font-bold text-xl">Speacialty</h3>
+          <h3 className="font-spartan font-bold text-xl">Specialty</h3>
           <p>{subjectArea.length > 0 ? <>{subjectArea.join(' | ')}</> : 'Please edit your profile'}</p>
           <h3 className="font-spartan font-bold text-xl text-center">
             Education & Experience
